google-cloud-run/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
Flask_Cors
Flask-SocketIO